+++ apps/web/src/app/(auth)/subscription/page.tsx
@@ -7,15 +7,50 @@ import HTTPMethods from "@/constants/http-methods"
 import Suspense from "@/components/suspense"
 import Loading from "@/components/loading"
 import { TierCardComponent } from "@/components/tiercard"
+import axios from "axios"
+import { toast } from "@/components/ui/use-toast"
+import { ToastAction } from "@/components/ui/toast"
+import { Button } from "@/components/ui/button"
+import { useContext, useState } from "react"
+import { GlobalContext } from "@/context/providers/globalstate.provider"
 
 export default function Page() {
+  const [{ userState }] = useContext(GlobalContext)
   const pricing = useQuery(["pricing"], endPoints.getSubscriptionConfig, HTTPMethods.GET)
+  const [selectedPlan, setSelectedPlan] = useState("premium")
+
+  const handlePayment = async () => {
+    if (userState.hasActiveSubscription) {
+      toast({
+        title: "Notification",
+        description: <p className="text-neutral-600">You already have an active subscription</p>,
+        action: <ToastAction altText="Goto schedule to undo">Okay</ToastAction>
+      })
+    }
+
+    else {
+      try {
+        const response = await axios.post(endPoints.createCheckoutSession, { selectedPlan: selectedPlan })
+        window.location = response.data.redirectUrl
+      }
+
+      catch (error) {
+        toast({
+          title: "Notification",
+          description: <p className="text-neutral-600">Error creating checkout session</p>,
+          action: <ToastAction altText="Goto schedule to undo">Okay</ToastAction>
+        })
+      }
+    }
+  }
 
   const renderPricing = pricing?.data?.map((pricing: any) => {
     return (
       <li className="flex" key={pricing.planName}>
         <TierCardComponent
           className={cn(pricing.length === 1 && "xl-col-span-2 xl:col-start-2")}
+          handleClick={(planName: string): void => setSelectedPlan(planName)}
+          isSelected={pricing.planName === selectedPlan}
           {...pricing}
         />
       </li>
@@ -40,6 +75,13 @@ export default function Page() {
             <ul className={cn("mx-auto grid max-w-md grid-cols-1 gap-8 md:max-w-2xl md:grid-cols-2 lg:max-w-4xl xl:mx-0 xl:max-w-none xl:grid-cols-3", pricing?.data?.length > 3 && "2xl:grid-cols-4")}>
               {renderPricing}
             </ul>
+          </div>
+          <div className="flex items-center justify-center mt-6">
+            <Button onClick={handlePayment}>
+              <Suspense condition={selectedPlan !== "hobby"} fallback={<>Activate for free</>}>
+                Pay & Subscribe
+              </Suspense>
+            </Button>
           </div>
         </section>
       </div>